Combined event wrappers; got controlled component calling back to consumer when it would normally close
